// frontend/src/app/generated/client/models/mailbox-message-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MailboxMessageDto {
  bodyKey: string;
  category: string;
  createdAt: string;
  id: string;
  isRead: boolean;
  link: (string | null);
  parameters: {
[key: string]: string;
};
  readAtUtc: (string | null);
  titleKey: string;
}